OCSADV-255 Move GemsUtils4Java.optimizeRadiusConstraint to GemsVoTableCatalog #time 20m
