+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/gems/GemsUtils4Java.scala
@@ -1,16 +1,13 @@
 package edu.gemini.ags.gems
 
-import edu.gemini.catalog.api.{SaturationConstraint, RadiusConstraint, MagnitudeConstraints}
+import edu.gemini.catalog.api.MagnitudeConstraints
 import edu.gemini.ags.impl._
-import edu.gemini.shared.util.immutable
 import edu.gemini.spModel.core.Target.SiderealTarget
 import edu.gemini.spModel.core._
 import edu.gemini.spModel.target.SPTarget
-import edu.gemini.shared.util.immutable.ScalaConverters._
 import edu.gemini.shared.skyobject
 import edu.gemini.skycalc
 import edu.gemini.spModel.target.system.HmsDegTarget
-import scala.math._
 import scala.collection.JavaConverters._
 
 import scalaz._
@@ -23,26 +20,6 @@ object GemsUtils4Java {
   // Returns true if the target magnitude is within the given limits
   def containsMagnitudeInLimits(target: SiderealTarget, magLimits: MagnitudeConstraints): Boolean =
     target.magnitudeIn(magLimits.band).map(m => magLimits.contains(m)).getOrElse(true)
-
-  def mapMagnitudes(constraint: Option[SaturationConstraint], mapOp: immutable.MapOp[SaturationConstraint, java.lang.Double]): immutable.Option[java.lang.Double] =
-    constraint.map(m => mapOp.apply(m)).asGeminiOpt
-
-  // Combines multiple radius limits into one
-  def optimizeRadiusConstraint(criterList: java.util.List[GemsCatalogSearchCriterion]): RadiusConstraint = {
-    val result = criterList.asScala.foldLeft((Double.MinValue, Double.MaxValue)) { (prev, current) =>
-      val c = current.criterion
-      val radiusConstraint = c.adjustedLimits
-      val maxLimit = radiusConstraint.maxLimit
-      val correctedMax = (c.offset |@| c.posAngle) { (o, _) =>
-          // If an offset and pos angle were defined, normally an adjusted base position
-          // would be used, however since we are merging queries here, use the original
-          // base position and adjust the radius limits
-          maxLimit + o.distance
-        } | maxLimit
-      (max(correctedMax.toDegrees, prev._1), min(radiusConstraint.minLimit.toDegrees, prev._2))
-    }
-    RadiusConstraint.between(Angle.fromDegrees(result._1), Angle.fromDegrees(result._2))
-  }
 
   /**
    * Sorts the targets list, putting the brightest stars first and returns the sorted array.
