+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/gems/GemsVoTableCatalog.scala
@@ -3,7 +3,7 @@ package edu.gemini.ags.gems
 import edu.gemini.catalog.api._
 import edu.gemini.catalog.votable.{RemoteBackend, VoTableBackend, CatalogException, VoTableClient}
 import edu.gemini.spModel.core.Target.SiderealTarget
-import edu.gemini.spModel.core.{Magnitude, MagnitudeBand, Coordinates}
+import edu.gemini.spModel.core.{Angle, Magnitude, MagnitudeBand, Coordinates}
 import edu.gemini.spModel.gemini.gems.GemsInstrument
 import edu.gemini.spModel.obs.context.ObsContext
 
@@ -11,6 +11,7 @@ import scala.concurrent.{Await, Future}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
 import scala.collection.JavaConverters._
+import scala.math._
 
 import scalaz._
 import Scalaz._
@@ -130,8 +131,25 @@ case class GemsVoTableCatalog(backend: VoTableBackend = RemoteBackend) {
   protected [gems] def getRadiusLimits(inst: GemsInstrument, criterions: List[GemsCatalogSearchCriterion]): List[RadiusConstraint] = {
     inst match {
       case GemsInstrument.flamingos2 => criterions.map(_.criterion.adjustedLimits)
-      case _                         => List(GemsUtils4Java.optimizeRadiusConstraint(criterions.asJava))
+      case _                         => List(optimizeRadiusConstraint(criterions.asJava))
     }
+  }
+
+  // Combines multiple radius limits into one
+  protected [gems] def optimizeRadiusConstraint(criterList: java.util.List[GemsCatalogSearchCriterion]): RadiusConstraint = {
+    val result = criterList.asScala.foldLeft((Double.MinValue, Double.MaxValue)) { (prev, current) =>
+      val c = current.criterion
+      val radiusConstraint = c.adjustedLimits
+      val maxLimit = radiusConstraint.maxLimit
+      val correctedMax = (c.offset |@| c.posAngle) { (o, _) =>
+          // If an offset and pos angle were defined, normally an adjusted base position
+          // would be used, however since we are merging queries here, use the original
+          // base position and adjust the radius limits
+          maxLimit + o.distance
+        } | maxLimit
+      (max(correctedMax.toDegrees, prev._1), min(radiusConstraint.minLimit.toDegrees, prev._2))
+    }
+    RadiusConstraint.between(Angle.fromDegrees(result._1), Angle.fromDegrees(result._2))
   }
 
   // Sets the min/max magnitude limits in the given query arguments
