+++ bundle/edu.gemini.ags/src/test/scala/edu/gemini/ags/gems/GemsVoTableCatalogSpec.scala
@@ -5,7 +5,7 @@ import edu.gemini.catalog.votable.TestVoTableBackend
 import edu.gemini.spModel.gemini.gems.Canopus.Wfs
 import org.specs2.time.NoTimeConversions
 import scala.concurrent.duration._
-import edu.gemini.shared.util.immutable.None
+import edu.gemini.shared.util.immutable.{None => JNone}
 import edu.gemini.spModel.core._
 import edu.gemini.spModel.core.AngleSyntax._
 import edu.gemini.spModel.gemini.gems.GemsInstrument
@@ -16,6 +16,7 @@ import edu.gemini.spModel.obs.context.ObsContext
 import edu.gemini.spModel.target.SPTarget
 import edu.gemini.spModel.target.env.TargetEnvironment
 import edu.gemini.spModel.telescope.IssPort
+import AlmostEqual.AlmostEqualOps
 
 import org.specs2.mutable.Specification
 
@@ -26,6 +27,7 @@ import scala.concurrent.Await
 import scala.collection.JavaConverters._
 
 class GemsVoTableCatalogSpec extends Specification with NoTimeConversions {
+  val magnitudeConstraints = MagnitudeConstraints(MagnitudeBand.J, FaintnessConstraint(10.0), SaturationConstraint(2.0).some)
 
   "GemsVoTableCatalog" should {
     "support executing queries" in {
@@ -36,7 +38,7 @@ class GemsVoTableCatalogSpec extends Specification with NoTimeConversions {
       val inst = new Gsaoi
       inst.setPosAngle(0.0)
       inst.setIssPort(IssPort.SIDE_LOOKING)
-      val ctx = ObsContext.create(env, inst, None.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
+      val ctx = ObsContext.create(env, inst, JNone.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
       val base = Coordinates(RightAscension.fromAngle(ra), Declination.fromAngle(dec).getOrElse(Declination.zero))
       val opticalCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
       val nirCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
@@ -63,7 +65,7 @@ class GemsVoTableCatalogSpec extends Specification with NoTimeConversions {
       val inst = new Gsaoi
       inst.setPosAngle(0.0)
       inst.setIssPort(IssPort.SIDE_LOOKING)
-      val ctx = ObsContext.create(env, inst, None.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
+      val ctx = ObsContext.create(env, inst, JNone.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
       val opticalCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
       val nirCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
       val instrument = GemsInstrument.gsaoi
@@ -85,7 +87,7 @@ class GemsVoTableCatalogSpec extends Specification with NoTimeConversions {
       val inst = new Gsaoi
       inst.setPosAngle(0.0)
       inst.setIssPort(IssPort.SIDE_LOOKING)
-      val ctx = ObsContext.create(env, inst, None.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
+      val ctx = ObsContext.create(env, inst, JNone.instance[Site], SPSiteQuality.Conditions.BEST, null, null)
       val opticalCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
       val nirCatalog = GemsGuideStarSearchOptions.DEFAULT_CATALOG
       val instrument = GemsInstrument.gsaoi
@@ -99,6 +101,59 @@ class GemsVoTableCatalogSpec extends Specification with NoTimeConversions {
       results should be size 2
       results(0) should beEqualTo(MagnitudeConstraints(MagnitudeBand.R, FaintnessConstraint(16), Some(SaturationConstraint(8.5))))
       results(1) should beEqualTo(MagnitudeConstraints(MagnitudeBand.H, FaintnessConstraint(14.5), Some(SaturationConstraint(7.3))))
+    }
+    "preserve the radius constraint for a single item without offsets" in {
+      val catalog = GemsVoTableCatalog(TestVoTableBackend(""))
+      val key = new GemsCatalogSearchKey(GemsGuideStarType.flexure, GsaoiOdgw.Group.instance)
+      val radiusConstraint = RadiusConstraint.between(Angle.fromArcmin(10.0), Angle.fromArcmin(2.0))
+      val criterion = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint, None, None)
+
+      val s = new GemsCatalogSearchCriterion(key, criterion)
+      catalog.optimizeRadiusConstraint(List(s).asJava).maxLimit ~= radiusConstraint.maxLimit
+      catalog.optimizeRadiusConstraint(List(s).asJava).minLimit ~= radiusConstraint.minLimit
+    }
+    "offset the radius constraint for a single item with offsets" in {
+      val catalog = GemsVoTableCatalog(TestVoTableBackend(""))
+      val key = new GemsCatalogSearchKey(GemsGuideStarType.flexure, GsaoiOdgw.Group.instance)
+      val radiusConstraint = RadiusConstraint.between(Angle.fromArcmin(10.0), Angle.fromArcmin(2.0))
+      val offset = Offset(Angle.fromArcmin(3), Angle.fromArcmin(4)).some
+      val posAngle = Angle.fromArcmin(3).some
+      val criterion = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint, offset, posAngle)
+
+      val s = new GemsCatalogSearchCriterion(key, criterion)
+      catalog.optimizeRadiusConstraint(List(s).asJava).maxLimit ~= radiusConstraint.maxLimit + Angle.fromArcmin(5)
+      catalog.optimizeRadiusConstraint(List(s).asJava).minLimit ~= radiusConstraint.minLimit
+    }
+    "find the max and min for a list of radius constraint without offsets" in {
+      val catalog = GemsVoTableCatalog(TestVoTableBackend(""))
+      val key = new GemsCatalogSearchKey(GemsGuideStarType.flexure, GsaoiOdgw.Group.instance)
+      val radiusConstraint1 = RadiusConstraint.between(Angle.fromArcmin(10.0), Angle.fromArcmin(2.0))
+      val radiusConstraint2 = RadiusConstraint.between(Angle.fromArcmin(15.0), Angle.fromArcmin(3.0))
+      val criterion1 = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint1, None, None)
+      val criterion2 = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint2, None, None)
+
+      val s1 = new GemsCatalogSearchCriterion(key, criterion1)
+      val s2 = new GemsCatalogSearchCriterion(key, criterion2)
+      catalog.optimizeRadiusConstraint(List(s1, s2).asJava).maxLimit ~= Angle.fromArcmin(15.0)
+      catalog.optimizeRadiusConstraint(List(s1, s2).asJava).minLimit ~= Angle.fromArcmin(2.0)
+    }
+    "find the max and min for a list of radius constraints with offsets" in {
+      val catalog = GemsVoTableCatalog(TestVoTableBackend(""))
+      val key = new GemsCatalogSearchKey(GemsGuideStarType.flexure, GsaoiOdgw.Group.instance)
+      val radiusConstraint1 = RadiusConstraint.between(Angle.fromArcmin(10.0), Angle.fromArcmin(2.0))
+      val radiusConstraint2 = RadiusConstraint.between(Angle.fromArcmin(15.0), Angle.fromArcmin(3.0))
+
+      val offset1 = Offset(Angle.fromArcmin(3), Angle.fromArcmin(4)).some
+      val offset2 = Offset(Angle.fromArcmin(5), Angle.fromArcmin(12)).some
+      val posAngle = Angle.fromArcmin(3).some
+      val criterion1 = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint1, offset1, posAngle)
+      val criterion2 = CatalogSearchCriterion("test", magnitudeConstraints.some, radiusConstraint2, offset2, posAngle)
+
+      val s1 = new GemsCatalogSearchCriterion(key, criterion1)
+      val s2 = new GemsCatalogSearchCriterion(key, criterion2)
+      // Gets the offset from the largest offset distance (offset2 in this case)
+      catalog.optimizeRadiusConstraint(List(s1, s2).asJava).maxLimit ~= (Angle.fromArcmin(15.0) + Angle.fromArcmin(13))
+      catalog.optimizeRadiusConstraint(List(s1, s2).asJava).minLimit ~= Angle.fromArcmin(2.0)
     }
   }
 }
